refactor: move MessageDocumentProps
move from Doppler.PushContact to Doppler.PushContact.Models
